counting items from backend

+++ final-capstone/java/src/main/java/com/techelevator/dao/ItemDao.java
@@ -19,7 +19,7 @@ public interface ItemDao {
 
    Item getItemByListId (String item);
 
-   int getItemCount (int listId);
+   long getItemCount (long listId);
 
    void create(String name, int itemAmount, int itemId);
